### 2025-10-03 #9137 @danielskatz
I am referring to the README, but more or less what's in the paper would probably be enough

> And in the README, also say something about what HELICS is please
