### 2025-10-03 b98cf6b
Update content of README and JOSS paper
