+++ docs/joss/paper.bib
@@ -31,7 +31,8 @@
 	month = {Feb},
 	pages = {1-6},
 	title = {Community-Based Transactive Coordination Mechanism for Enabling Grid-Edge Systems},
-	year = {2024}
+	year = {2024},
+	url = {https://ieeexplore.ieee.org/document/10472237}
 }
 
 
