+++ run/python/data_example_json_csv/cst_post_processing_example.py
@@ -1,0 +1,374 @@
+from cosim_toolbox.dbms import create_metadata_manager
+from prettytable import PrettyTable
+from pprint import pprint
+from cosim_toolbox.dbms import create_timeseries_manager
+from cosim_toolbox.dbms import TSRecord
+from cosim_toolbox.dbms.csv_timeseries import (
+    CSVTimeSeriesWriter,
+    CSVTimeSeriesReader,
+    CSVTimeSeriesManager,
+    _CSVHelper,
+)
+from datetime import datetime
+import time as t
+import pandas as pd
+import os
+
+
+class post_process_example:
+
+    def create_csv_files_meta_data(selfself,temp_dir):
+        """Create the metadata database and load it with data
+
+        Args:
+            temp_dir (str): directory path to the folder that the metadata will be written to
+        """
+        # Create metadata dictionaries for the data contained in the csv files
+        scenario1 = {
+            "analysis": "rooftop solar",
+            "federation": "rooftop solar federation",
+            "start_time": "2025-01-01T00:00:00",
+            "stop_time": "2025-12-31T23:00:00",
+            "docker": False,
+            "solar tilt angle": 20,
+            "azimuth": 90,
+            "units": "kW"
+        }
+        scenario2 = {
+            "analysis": "rooftop solar",
+            "federation": "rooftop solar federation",
+            "start_time": "2025-01-01T00:00:00",
+            "stop_time": "2025-12-31T23:00:00",
+            "docker": False,
+            "solar tilt angle": 20,
+            "azimuth": 180,
+            "units": "kW"
+        }
+        scenario3 = {
+            "analysis": "rooftop solar",
+            "federation": "rooftop solar federation",
+            "start_time": "2025-01-01T00:00:00",
+            "stop_time": "2025-12-31T23:00:00",
+            "docker": False,
+            "solar tilt angle": 20,
+            "azimuth": 270,
+            "units": "kW"
+        }
+        scenario4 = {
+            "analysis": "rooftop solar",
+            "federation": "rooftop solar federation",
+            "start_time": "2025-01-01T00:00:00",
+            "stop_time": "2025-12-31T23:00:00",
+            "docker": False,
+            "solar tilt angle": 30,
+            "azimuth": 90,
+            "units": "kW"
+        }
+        scenario5 = {
+            "analysis": "rooftop solar",
+            "federation": "rooftop solar federation",
+            "start_time": "2025-01-01T00:00:00",
+            "stop_time": "2025-12-31T23:00:00",
+            "docker": False,
+            "solar tilt angle": 30,
+            "azimuth": 180,
+            "units": "kW"
+        }
+        scenario6 = {
+            "analysis": "rooftop solar",
+            "federation": "rooftop solar federation",
+            "start_time": "2025-01-01T00:00:00",
+            "stop_time": "2025-12-31T23:00:00",
+            "docker": False,
+            "solar tilt angle": 30,
+            "azimuth": 270,
+            "units": "kW"
+        }
+        scenario7 = {
+            "analysis": "rooftop solar",
+            "federation": "rooftop solar federation",
+            "start_time": "2025-01-01T00:00:00",
+            "stop_time": "2025-12-31T23:00:00",
+            "docker": False,
+            "solar tilt angle": 40,
+            "azimuth": 90,
+            "units": "kW"
+        }
+        scenario8 = {
+            "analysis": "rooftop solar",
+            "federation": "rooftop solar federation",
+            "start_time": "2025-01-01T00:00:00",
+            "stop_time": "2025-12-31T23:00:00",
+            "docker": False,
+            "solar tilt angle": 40,
+            "azimuth": 180,
+            "units": "kW"
+        }
[…]
