### 2025-12-14 pnnl/cst#50 @Puneetha-Ramachandra
[joss-review] Suggestions for "State of the Field", Audience Definition, and Minor Fixes
openjournals/joss-reviews#9540

1. There are a few typos that need correction in the paper:
    * "exhanges" should be "exchanges"
    * "sucessfully" should be "successfully"
    * "evaulating" and "evaluting" should be "evaluating"
    * "Particuarly" should be "Particularly"
    * "provide a a web interface" has a duplicate "a"
    * "Acknowlegements" is misspelled (missing 'd')

2. State of the Field: A brief section comparing CST to similar Python frameworks (like Mosaik) or native HELICS CLI tools would greatly help readers understand its unique value. If no direct alternative exists, explicitly stating that CST fills a gap left by native HELICS tools would be equally valuable.

3. Citations: The paper lacks a citation or link for "Modelon" and references for key infrastructure tools (Docker, Postgres, MongoDB). Adding footnotes or citations for these dependencies would improve the completeness of the software attribution.

4. Target Audience: It can be inferred from the "Statement of Need" that the software is intended to help users of HELICS, but the text does not explicitly identify the specific target audience. Adding a sentence to clarify the intended user base would strengthen the section.
