### 2025-12-14 pnnl/cst#51 @Puneetha-Ramachandra
MongoDB Initialization Failure: sudo Permission Issue
openjournals/joss-reviews#9540

I encountered an issue where the MongoDB services failed to initialize the worker user, leading to a persistent `Authentication failed error (Code 18)` when running federates_config.py.

The root cause was a permissions error during the Docker startup process, which occurred specifically because I was running docker compose up with sudo.

The Docker logs clearly showed a failure to access the initialization script: Error:` EACCES: permission denied, open '/docker-entrypoint-initdb.d/init-mongo.js'`

The user defined in the init-mongo.js script (worker) was never created in the database because the MongoDB entrypoint script—which runs as an internal container user—did not have the necessary execute permission to run the mounted host file. This issue is specific to sudo users because of the way host file permissions are restricted.

The issue was resolved by explicitly granting read and execute permissions to the init-mongo.js file on the host before running the Docker stack:
```
chmod a+rwx /path/to/init-mongo.js
```

The file permissions for init-mongo.js should be explicitly set (e.g., using chmod a+rx) during the setup phase to ensure the MongoDB entrypoint can execute it, regardless of the user's host permissions or the use of sudo.
