### 2025-12-29 #9540 @trevorhardy
> [@prashjha](https://github.com/prashjha), I'm waiting on the authors to resolve [pnnl/cst#50](https://github.com/pnnl/cst/issues/50) and [pnnl/cst#51](https://github.com/pnnl/cst/issues/51). I'll post here once those are addressed and my checklist is completed.

I've tasked the development team to look into these and I'm optimistic they will be resolved in the coming week or so. Thanks for the helpful feedback.
