### 2026-01-13 0d24735
Misc document comments, fix issue #51, fix pypi worklflow
