+++ docs/conf.py
@@ -32,14 +32,14 @@ from recommonmark.parser import CommonMarkParser
 
 # If your documentation needs a minimal Sphinx version, state it here.
 #
-needs_sphinx = '8.2.0'
+needs_sphinx = '8.2.3'
 
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
 
 numfig = True
-
+suppress_warnings = ["myst.xref_missing"]
 extensions = ['sphinx.ext.autodoc',
               'sphinx.ext.imgmath',
               'sphinx.ext.githubpages',
