+++ run/data_management/example_timeseries_postgres.py
@@ -88,9 +88,8 @@ records = [
     ),
 ]
 
-with create_timeseries_manager(
-    backend="postgres", analysis_name="example_data"
-) as mgr:
+with create_timeseries_manager(backend="postgres", analysis_name="example_data") as mgr:
+    mgr.connect()
     mgr.delete_scenario_data(scenario_name="test_scenario")
     mgr.write_records(records=records)
     print(f"scenarios: {mgr.list_scenarios()}")
