+++ docs/joss/paper.md
@@ -31,19 +31,21 @@ bibliography: paper.bib
 ---
 
 # Summary
-Co-simulation is a technique for creating larger and more complex time-series system models by facilitating data exchange between two or more models during runtime. This data exchange allows boundary states in models which may have been defined through static values or historical values to be dynamically generated by models in other simulator instances. This dynamic data exchange allows for system models that transcend individual simulation tools, modeled domains, and modeled subsystems. FMI [@fmi_github], particularly as implemented in Modelon, is an example of a popular co-simulation platform.
+Co-simulation is a technique for creating larger and more complex time-series system models by facilitating data exchange between two or more models during runtime. This data exchange allows boundary states in models which may have been defined through static values or historical values to be dynamically generated by models in other simulator instances. This dynamic data exchange allows for system models that transcend individual simulation tools, modeled domains, and modeled subsystems. FMI [@fmi_github], particularly as implemented in Modelica [@modelica] and Mosaik [@app9050923], are examples of a popular co-simulation platforms.
 
 HELICS [@Hardy:2024aa] is the US Department of Energy's publicly available co-simulation platform and has been under development since 2017. HELICS provides the necessary simulation time management and facilitates the data exhanges between simulator instances (called "federates") and has been sucessfully used in a number of published studies such as evaluating transactive retail market in power systems[@Theisen:2024aa], evaulating the roles of distributed power inverters on power system frequency response[@Bharati:2023aa], evaluting the impact of electric vehicle charging in a metropolitan area [@Panossian:2023aa],  and cyber-security analysis of power system[@Lardier:2020aa]. Due to its origins as a power system tool, many of the existing studies are connected to that domain but HELICS itself is domain agnostic and is suitable for use in any time-series simulation. 
 
 # Statement of Need
-CoSim Toolbox (CST) is built on top of HELICS and has the goal of making any HELICS-based co-simulations easier to manage. Based on years of experience using HELICS, we have found a number of pain points around data and metadata management and HELICS federate creation that CoSim Toolbox is designed to alleviate. Specifically, we have integrated the following tools and code to make using HELICS easier.
+CoSim Toolbox (CST) is built on top of HELICS and has the goal of making any HELICS-based co-simulations easier to manage and HELICS users are the target audience for CST. Based on years of experience using HELICS, we have found a number of pain points around data and metadata management and HELICS federate creation that CoSim Toolbox is designed to alleviate. Specifically, we have integrated the following tools and code to make using HELICS easier.
 
 - **Docker** - Docker containers are used to distribute HELICS, the CST API as well as any CST-enabled simulation tools. This allows the use of CST on all Linux, MacOS, and Windows while minimizing the maintenance of these codebases.
 - **Generic Federate class** - To facilitate writing new federates more easily, CST provides a generic federate class for use when writing new Python federates. This class wraps the HELICS APIs and facilitates data exchanges using just Python data structures. By minimizing the amount of knowledge of HELICS APIs, we have seen those with no knowledge of HELICS be able to more easily implement a HELICS-based co-simulation using this CST federate class.
 - **Time-Series Data Collection** - A Postgres database is used to collect time-series data being generated by every federate. A data collection federate has been written using the CST APIs to collect data generated by the HELICS-enabled data exchanges and automatically commit it to the Postgres database. This allows data to be stored and thus accessed in a central location, making it easier to monitor and post-process the co-simulation data. The Postgres database is implemented as a Docker container along with pgAdmin to provide a web interface for inspecting the Postgres database.
[…]
