### 2026-01-13 pnnl/cst#50 @trevorhardy
2 - Added brief paragraph comparing to existing tooling in Mosaik. I wasn't even citing Mosaik before which was a significant oversight on my part; I'm glad this was caught in review.
3 - I realized the citation should have been to the software Modelica (not the company, Modelon) and have added a reference.
4 - Explicit statement of target audience added.

Change made and in PR #54
