### 2026-01-26 77bd3e2
add catalog-info.yaml file for forge catalog
