### 2026-02-07 #9540 @prashjha
Hi @trevorhardy, please can you take a look at the issue in compiling pdf?

Also, please archive the code release on Zenodo (if not already done) and provide the archive reference so I can associate it with your JOSS submission? Also, please make sure that the title of the Zenodo archive matches the title of this JOSS article.

If you have released a new version of the code release and would like to update the version here, let me know.

I will review and share the suggestions for the Joss draft (if any) once you fix the pdf compilation issue as a GitHub issue soon.
