### 2026-02-12 #9540 @trevorhardy
@prashjha, PDF is compiling now. I'm going to get to work on getting the code submitted to Zenodo.

(FYI, the build failed unless working from the original "joss" branch, even when I told it to switch branches to "main" as we had deleted "joss". [It was clear from the  log that build was still trying to use "joss"](https://github.com/openjournals/joss-papers/actions/runs/21963695500/job/63447705999#step:4:103). I'm assuming this isn't something you are responsible for but maybe you know who needs to look into this?)
