### 2026-02-12 #9540 @trevorhardy
@prashjha, the released version of the code used on Zenodo (and what I assume we want to reference here) is v1.0.1. This contains the latest draft of the JOSS paper.
