### 2026-02-17 88ff7f6
JOSS reviews: minor edits
Related to openjournals/joss-reviews#9540
