### 2026-02-21 #9540 @trevorhardy
@prashjha, had some extra braces; fix committed to the "joss" branch.
