### 2026-02-21 #9540 @prashjha
Hi @trevorhardy, you earlier mentioned 'main' branch for pdf compilation. Do you want me to change it the branch here to 'joss'?
