### 2026-03-08 #9540 @prashjha
Hello @trevorhardy, I changed the branch to joss for paper compilation. Looking at the PDF, the draft seems to be old and does not include the changes I made to your repository. Please can you check whether the draft in joss branch is up to date? If not, can you update it so the pdf can be regenerated here?
