+++ docs/joss/paper.bib
@@ -1,7 +1,7 @@
 
 @article{Hardy:2024aa,
 	author = {Hardy, Trevor D. and Palmintier, Bryan and Top, Philip L. and Krishnamurthy, Dheepak and Fuller, Jason C.},
-	doi = {{10.1109/ACCESS.2024.3363615}},
+	doi = {10.1109/ACCESS.2024.3363615},
 	journal = {{IEEE Access}},
 	keywords = {PNNL,Natural gas;Renewable energy sources;Government;Transactive energy;Scalability;Power system stability;Power system dynamics;Simulation;Power system analysis computing;power system simulation;HELICS;co-simulation;natural gas;transportation;multi-energy analysis;multi-domain analysis;energy system analysis},
 	month = {1},
@@ -32,7 +32,7 @@
 	abstract = {The changing landscape of the electricity industry, characterized by a surge in distributed energy resources (DERs) and proactive customers, necessitates practical solutions for coor-dinated operation especially at the distribution-level. This paper introduces a community-based trans active coordination mecha-nism designed to incentivize customers for providing localized and system-level services reflected through real-time prices. The work presents a bidding approach for communities, representing collectives of customers, to formulate their price-responsiveness for retail energy coordination, thereby emphasizing a community-centric model. By sending bidding curves to a third-party, the mechanism enables customers with DER assets to actively participate in localized coordination with the Load Serving Entity (LSE), thereby supplementing each other's and even the utility's needs through a shared energy economy. The proposed transactive mechanism is implemented leveraging a co-simulation framework that integrates a distribution grid simulator with control agents for performance evaluation. Simulation-based evaluation on a real distribution system use-case, in collaboration with a local utility, demonstrate the potential of the mechanism to reduce costs by 12% for communities with DERs like solar photovoltaic (PV) and battery energy storage systems (BESS).},
 	author = {Theisen, John R. and Bose, Anjan and Mukherjee, Monish and Burgess, Dan and Wilhelm, Kenneth and Diedesch, Michael},
 	booktitle = {2024 IEEE Texas Power and Energy Conference (TPEC)},
-	doi = {{10.1109/TPEC60005.2024.10472237}},
+	doi = {10.1109/TPEC60005.2024.10472237},
 	keywords = {Photovoltaic systems;Performance evaluation;Reactive power;Costs;Simulation;Pricing;Real-time systems;Community-based coordination;grid-edge sys-tems;retail market;shared energy economy;transactive energy; co-simulation; HELICS},
 	month = {Feb},
 	pages = {1-6},
@@ -45,7 +45,7 @@
 @article{Bharati:2023aa,
[…]
