### 2026-03-09 #9540 @trevorhardy
@prashjha; good catch. I thought cleaning up the .bib formatting in the Github web editor would be fastest but that wasn't the case. Ok, I've merged `main` back into `joss` and then merged `joss` back into `main`; both branches are identical now.

Sorry for making this complicated.
