### 2026-03-10 1bd624b
minor changes in paper.md
