+++ docs/joss/paper.md
@@ -32,19 +32,19 @@ bibliography: paper.bib
 ---
 
 # Summary
-Co-simulation is a technique for creating larger and more complex time-series system models by facilitating data exchange between two or more models during runtime. This data exchange allows boundary states in models which may have been defined through static values or historical values to be dynamically generated by models in other simulator instances. This dynamic data exchange allows for system models that transcend individual simulation tools, modeled domains, and modeled subsystems. FMI [@fmi_github], particularly as implemented in Modelica [@modelica] and Mosaik [@app9050923], are examples of popular co-simulation platforms.
+Co-simulation is a technique for creating larger and more complex time-series system models by facilitating data exchange between two or more models during runtime. This data exchange allows boundary states in models that may have been defined through static values or historical values to be dynamically generated by models in other simulator instances. This dynamic data exchange allows for system models that transcend individual simulation tools, modeled domains, and modeled subsystems. FMI [@fmi_github], particularly as implemented in Modelica [@modelica] and Mosaik [@app9050923], are examples of popular co-simulation platforms.
 
 HELICS [@Hardy:2024aa] is the US Department of Energy's publicly available co-simulation platform and has been under development since 2017. HELICS provides the necessary simulation time management and facilitates the data exchanges between simulator instances (called "federates"). HELICS has been successfully used in a number of published studies such as evaluating transactive retail market in power systems [@Theisen:2024aa], evaluating the roles of distributed power inverters on power system frequency response [@Bharati:2023aa], evaluating the impact of electric vehicle charging in a metropolitan area [@Panossian:2023aa],  and cyber-security analysis of power systems [@Lardier:2020aa]. Due to its origins as a power system tool, many of the existing studies are connected to that domain but HELICS itself is domain agnostic and is suitable for use in any time-series simulation. 
 
 # Statement of Need
-CoSim Toolbox (CST) is built on top of HELICS and has the goal of making any HELICS-based co-simulations easier to manage and HELICS users are the target audience for CST. Based on years of experience using HELICS, we have found a number of pain points around data and metadata management and HELICS federate creation that CoSim Toolbox is designed to alleviate. Specifically, we have integrated the following tools and code to make HELICS easier to use:
+CoSim Toolbox (CST) is built on top of HELICS and has the goal of making any HELICS-based co-simulation easier to manage; HELICS users are the target audience for CST. Based on years of experience using HELICS, we have found a number of pain points around data and metadata management and HELICS federate creation that CoSim Toolbox is designed to alleviate. Specifically, we have integrated the following tools and code to make HELICS easier to use:
 
-- **Docker** - Docker containers are used to distribute HELICS, the CST API as well as any CST-enabled simulation tools. This allows the use of CST on all Linux, macOS, and Windows while minimizing the maintenance of these codebases.
+- **Docker** - Docker containers are used to distribute HELICS, the CST API, as well as any CST-enabled simulation tools. This allows the use of CST on all Linux, macOS, and Windows systems while minimizing the maintenance of these codebases.
 - **Generic Federate class** - To facilitate writing new federates more easily, CST provides a generic federate class for use when writing new Python federates. This class wraps the HELICS APIs and facilitates data exchanges using just Python data structures. By minimizing the amount of knowledge of HELICS APIs, we have seen those with no knowledge of HELICS be able to more easily implement a HELICS-based co-simulation using this CST federate class.
[…]
