### 2026-03-10 85eeb5d
minor changes in paper.bib
