+++ docs/joss/paper.bib
@@ -6,15 +6,15 @@
 	keywords = {PNNL,Natural gas;Renewable energy sources;Government;Transactive energy;Scalability;Power system stability;Power system dynamics;Simulation;Power system analysis computing;power system simulation;HELICS;co-simulation;natural gas;transportation;multi-energy analysis;multi-domain analysis;energy system analysis},
 	month = {1},
 	pages = {{24325-24347}},
-	title = {{HELICS: A Co-Simulation Framework for Scalable Multi-Domain Modeling and Analysis}},
+	title = {{HELICS}: A Co-Simulation Framework for Scalable Multi-Domain Modeling and Analysis},
 	url = {https://ieeexplore.ieee.org/document/10424422},
 	volume = {12},
 	year = {2024},
 }
 
 @misc{fmi_github,
-  author = {Modelica Association Project FMI},
-  title = {fmi: Functional Mock-Up Interface},
+  author = {{Modelica Association Project FMI}},
+  title = {{f}mi: Functional Mock-Up Interface},
   year = {2025},
   publisher = {GitHub},
   journal = {GitHub repository},
@@ -22,7 +22,7 @@
 }
 
 @misc{modelica,
-  author = {Modelica Association},
+  author = {{Modelica Association}},
   title = {Modelica},
   url = {https://modelica.org/}
 }
@@ -68,7 +68,7 @@
 	journal = {Energies},
 	keywords = {HELICS, electric vehicle charging, electric vehicles},
 	number = {5},
-	title = {Architecture for Co-Simulation of Transportation and Distribution Systems with Electric Vehicle Charging at Scale in the San Francisco Bay Area},
+	title = {Architecture for Co-Simulation of Transportation and Distribution Systems with Electric Vehicle Charging at Scale in the {S}an {F}rancisco {B}ay {A}rea},
 	url = {https://www.mdpi.com/1996-1073/16/5/2189},
 	volume = {16},
 	year = {2023}
