### 2026-03-10 #9540 @danielskatz
👋 @trevorhardy - my suggested changed are in https://github.com/pnnl/cst/pull/59 - please merge this, or let me know what you disagree with, then we can continue to acceptance and publication
