### 2026-03-24 #9540 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
