### 2026-03-24 #9540 @danielskatz
Congratulations to @trevorhardy (Trevor Hardy) and co-authors on your JOSS publication!!

JOSS completely relies on volunteers - we couldn't do this without you
